% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_io.R
\name{create_114_sector}
\alias{create_114_sector}
\title{Create 114 sector input-output table.}
\source{
https://www.abs.gov.au/statistics/economy/national-accounts/australian-national-accounts-input-output-tables/latest-release
}
\usage{
create_114_sector(path = NULL)
}
\arguments{
\item{path}{path to a folder containing two spreadsheets downloaded from the Australian Bureau of Statistics.
NULL by default which uses the \code{industry_flows} data installed with the package.
\itemize{
\item{Table 5. Industry by industry flow table}{}
\item{Table 20. Employment by industry}{}
}}
}
\value{
Silently returns a list with the 114 sector input-output table ("flows"), and the path to Table 5 and Table 20.
}
\description{
Creates a 114 sector input-output table from the ABS National Accounts: Input-Output tables
Industry by industry flow table (Table 5.) and Employment by industry table (Table 20.). The 114 sector version
created is modified to enable easy creation of the 19 sector model used to create Regional Input-Ouput tables.
The current version of the data is installed with the package as \code{industry_flows}. A path can be specified to instead read Table 5
and Table 20 from a local folder. This can be useful if a different year input-output table is required.
}
\examples{
create_114_sector()
create_114_sector(path = "local_data")

}
