% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_national_io_table.R
\name{create_lq}
\alias{create_lq}
\title{Create Location Quotient model input output tables}
\usage{
create_lq(data, type, progress)
}
\arguments{
\item{data}{data.frame from get_data()}

\item{type}{character.}
}
\value{
a list of matrices
}
\description{
Create Location Quotient model input output tables
}
