% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employment.R
\name{get_local_employment}
\alias{get_local_employment}
\title{Local employment}
\usage{
get_local_employment(region, year, adjust = TRUE)
}
\arguments{
\item{region}{A Local Government Area (LGA). See: \code{get_available_regions()}}

\item{year}{The census year to draw the data from. Must be a census year (2011, 2016, 2021).}

\item{adjust}{logical. The default (TRUE) adjusts the data to remove Inadequately described and Not stated.}
}
\value{
A dataframe with the local employment of \code{region} by industry.
}
\description{
Return the local employment in a specified local government area. Local employment refers to
those who live and work in the region.
Data can adjusted to remove the Inadequately described and Not stated employment. The adjustment process
assumes that employment in Inadequately described and Not stated has the same distribution of employment by
industry as the reported data. For example, if local employment in Inadequately described totals 20,
and local employment in Agriculture, Forestry and Fishing in the region is 5\% of total local employment, then
5\% * 20 of the Inadequately described employment is proportioned to Agriculture, Forestry and Fishing.

Data is derived from the Australian Bureau of Statistics Census - Employment by industry by place of work by usual
residence for 2011, 2016 and 2021.
}
\examples{
get_local_employment("Adelaide", 2021)
# In the 2016 Census, the Adelaide LGA is called 'Adelaide (C)' so the following will not work.
\dontrun{
get_local_employment("Adelaide" 2016)
}
}
