% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact_analysis}
\alias{impact_analysis}
\title{Impact Analysis}
\usage{
impact_analysis(region, impacts)
}
\arguments{
\item{region}{Local Government Area}

\item{impacts}{A matrix with 19 rows and at least 1 column containing expenditure impacts for each of the 19 sectors.
The matrix must have rownames equal to the 19 industry sectors.}
}
\value{

}
\description{
Conduct an impact analysis assessment for a Local Government Area.
}
\examples{
\dontrun{
impacts <- matrix(data = 1, nrow = 19, ncol = 1, dimnames = list(anzsic_swap$name))
impact_analysis("Adelaide", impacts)
}
}
