% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plot.R
\name{impact_plot}
\alias{impact_plot}
\title{Plot Impact Analyses}
\usage{
impact_plot(data, type, indicator, background_colour = "grey")
}
\arguments{
\item{data}{Impact analysis data from \code{impact_analysis()}}

\item{type}{Line or column. Single year analyses always use column.}

\item{indicator}{One of "emp" for employment, "grp" for Gross Regional Product, or "output" for Output}

\item{background_colour}{Background colour of the graph.}
}
\value{

}
\description{
Plot Impact Analyses
}
\examples{

}
