% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{live_and_work}
\alias{live_and_work}
\title{Employment (by Place of Work, by Place of Residence) by Industry data}
\format{
\subsection{\code{live_and_work}}{

A data frame with 36,564 rows and 5 columns.
\describe{
\item{industry}{ANZSIC 19 Sector Industry}
\item{lga_ur}{Place (Local Government Area) of usual residency}
\item{lga_pow}{Place (Local Government Area) of work}
\item{employment}{Total employment}
\item{year}{Census year}
}
}
}
\usage{
live_and_work
}
\description{
Total employment by industry, by place of work and by place of usual residence (based on Local Government Area boundies)
for ABS Census years 2011, 2016 and 2021. From ABS TableBuilder Pro. Only includes employment where people live and work in the
same LGA.
}
\keyword{datasets}
