% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national_19}
\alias{national_19}
\title{Australian 19 Sector Input-Output Table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 27 columns.
}
\usage{
national_19
}
\description{
Australian 19 Sector Input-Output Table
}
\keyword{datasets}
