% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{LGA 2022 State Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 541 rows and 3 columns.
}
\usage{
regions
}
\description{
LGA 2022 State Names
}
\keyword{datasets}
