% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_transaction_tables.R
\name{rtt_basic}
\alias{rtt_basic}
\title{Create Regional Input-Output Tables}
\usage{
rtt_basic(data = get_data(2021), region, type = "household")
}
\arguments{
\item{data}{a list of data frames generated by \code{get_data}}

\item{region}{string. The region (Local Government Area) of the desired Input-Output Table.}

\item{type}{string. "basic" for a basic input-output table, or "household" to split Wages and Salaries and Household Consumption.}
}
\value{
Matrix
}
\description{
Creates a 19 Sector Input-Output table for a specified Local Government Area. The regional
Input-Output tables are created from the Australian 19 Sector Input-Output table using the location quotient method.
Two versions of the Input-Output table can be created. A basic version which does not split Wages and Salaries or
Household Consumption by Local and Regional employment, and a split version which does.

See the documentation for \code{get_data} for a description of the data required.
}
\examples{
rtt_basic(data = get_data(2021), region = "Adelaide", type = "household")
}
