% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_gos_coe}
\alias{state_gos_coe}
\title{State Compensation of employees}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 304 rows and 5 columns.
}
\usage{
state_gos_coe
}
\description{
State accounts data on compensation of employees and gross operating surplus.
This data is used to adjust state input-output tables
}
\keyword{datasets}
