% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_national_io_table.R
\name{update_national_io_table}
\alias{update_national_io_table}
\title{Update National Input Output Table}
\usage{
update_national_io_table(progress = TRUE)
}
\arguments{
\item{progress}{logical. TRUE to show a progress bar}
}
\value{
TRUE
}
\description{
Update National Input Output Table
}
