% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{work}
\alias{work}
\title{Employment (by Place of Work) by Industry data}
\format{
\subsection{\code{work}}{

A data frame with 37,312 rows and 4 columns.
\describe{
\item{lga_pow}{Place (Local Government Area) of work}
\item{industry}{ANZSIC 19 Sector Industry}
\item{year}{Census year}
\item{employment}{Total employment}
}
}
}
\usage{
work
}
\description{
Total employment by industry by place (local government area) of work for ABS Census years 2011
2016 and 2021. ABS TableBuilder Pro
}
\keyword{datasets}
