% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_io.R
\name{create_19_sector}
\alias{create_19_sector}
\title{Create 19 Sector Input-Output Table.}
\usage{
create_19_sector(check = FALSE)
}
\arguments{
\item{path}{path to a folder containing two spreadsheets downloaded from the Australian Bureau of Statistics.
NULL by default which uses the \code{industry_flows} data installed with the package.}
}
\value{
tibble of the industry-industry flows between 19 ANZSIC Sectors for Australia.
}
\description{
Creates a 19 Sector Input-Output Table by aggregating the 114 Sector Input-Output Table (see: \code{create_114_sector()})
}
\examples{
create_19_sector()
}
