% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lq_models}
\alias{lq_models}
\title{Regional Input Output Tables (Household Split)}
\format{
\subsection{\code{lq_models}}{

A list of 556 matrices with 29 rows and 27 columns.
}
}
\usage{
lq_models
}
\description{
Regional Input-Output tables derived using the location quotient method from the Australian 19 Sector Input-Output table.
These tables split household expenditure and wages and salaries by those who live and work in the region and outside the region.
}
\keyword{datasets}
